// src/app/api-generated/model/badgeResDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface BadgeResDto { 
    id?: number;
    userId?: number;
    type?: BadgeResDto.TypeEnum;
}
export namespace BadgeResDto {
    export const TypeEnum = {
        BEGINNER: 'BEGINNER',
        LEARNER: 'LEARNER',
        EXPLORER: 'EXPLORER',
        CONTRIBUTOR: 'CONTRIBUTOR',
        HELPER: 'HELPER',
        CONSISTENT: 'CONSISTENT',
        COLLABORATOR: 'COLLABORATOR',
        ACHIEVER: 'ACHIEVER',
        MENTOR: 'MENTOR',
        LEGEND: 'LEGEND'
    } as const;
    export type TypeEnum = typeof TypeEnum[keyof typeof TypeEnum];
}